requirements.txt
appdirs==1.4.4
asgiref==3.3.1
Authlib==0.15.3
bcrypt==3.2.0
beautifulsoup4==4.10.0
bs4==0.0.1
cachelib==0.4.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.3
cryptography==3.4.7
distlib==0.3.1
Django==3.1.5
dnspython==2.1.0
docutils==0.16
email-validator==1.1.3
filelock==3.0.12
Flask==2.0.2
Flask-Bcrypt==0.7.1
Flask-CKEditor==0.4.6
Flask-Login==0.5.0
Flask-MySQL==1.5.2
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
form==0.0.1
greenlet==1.1.2
gunicorn==20.1.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.2
Kivy==2.0.0
Kivy-examples==2.0.0
Kivy-Garden==0.1.4
MarkupSafe==2.0.1
mpmath==1.2.1
mysql-connector-python==8.0.23
Pillow==8.1.0
protobuf==3.15.6
pycparser==2.20
pygame==2.0.1
Pygments==2.8.0
PyMySQL==1.0.2
pytz==2020.5
requests==2.25.1
simple-salesforce==1.11.1
six==1.15.0
soupsieve==2.2.1
SQLAlchemy==1.4.26
sqlparse==0.4.1
sympy==1.9
urllib3==1.26.3
virtualenv==20.4.2
Werkzeug==2.0.2
WTForms==2.3.3
